# swagger.yml
swagger: "2.0"
info:
  title: Snake-Server API
  description: "Snake-Server"
  version: 4.0.0
  license:
    name: MIT
  contact:
    name: Ivan Pushkin
    url: https://twitter.com/IvanDeveloper
basePath: /api
schemes:
  - http
  - https
host: localhost:8080
produces:
  - application/json
consumes:
  - application/x-www-form-urlencoded
paths:
...
  /games/{id}:
    get:
      description: Get game by identificator
      parameters:
        - name: id
          in: path
          type: integer
          format: int32
          required: true
          description: Game identificator
      responses:
        200:
          description: Returns game object
          schema:
            $ref: '#/definitions/Game'
        400:
          description: Invalid parameters
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Game not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: Delete game by identificator
      parameters:
        - name: id
          in: path
          type: integer
          format: int32
          required: true
          description: Game identificator
      responses:
        200:
          description: Returns identificator of deleted game
          schema:
            $ref: '#/definitions/Deleted'
        400:
          description: Invalid parameters
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Game not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
        503:
          description: Service unavailable, cannot delete game right away
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Game:
    type: object
    description: Object contains information about game
    required:
      - id
      - limit
      - count
      - width
      - height
    properties:
      id:
        description: Game identificator
        type: integer
        format: int32
      limit:
        description: Game players limit
        type: integer
        format: int32
      count:
        description: Current players count in game
        type: integer
        format: int32
      width:
        description: Map width
        type: integer
        format: int32
      height:
        description: Map height
        type: integer
        format: int32
...
  Error:
    type: object
    description: Contains code of error and short description text
    required:
      - code
      - text
    properties:
      code:
        description: Error code
        type: integer
        format: int32
      text:
        description: Short error description
        type: string
  Deleted:
    type: object
    description: Object contains information about deleted game
    required:
      - id
    properties:
      id:
        description: Game identificator
        type: integer
        format: int32